Fixed extrabuttons handling
The index for the extrabuttons is off by 2 from the other buttons. Also
it looks like the third one is a special command to trigger sampling the
device, so I put in a skeleton for that.

I also added displaying the error that happens on eventaction otherwise
you don't see that it can't talk to the device.