Implemented bare bones of Aillio data
This handles just the very basic stuff for the bullet. It can get the
bean temperature and the drum temperature. I'm using the drum
temperature for the ET because. Not sure if that's correct or not.

Also, I think I've got too much polling. The GUI is pretty slow while
the thing is running. I'm using the asynchronous API with usb1, but I've
got a tight loop waiting for messages to finish. I'm doing this because
the way the bullet works is it doesn't give you the latest stats when
you query it. It has a buffer of about 15 stored stats and it always
gives you the oldest one that it hasn't sent yet.

So what I've been doing is I keep requesting stats until it says I've
got the latest one. With an interval of 2 seconds that means every time
Artisan asks the bullet for stats, I'm sending and waiting for about 8
messages (4 stats and 4 status). I don't understand enough about how
Artisan wants to work with devices to decide how I should fix this up.