Reorged Bullet code
I'd like to try and keep it to one class per file and I wanted to add an
exception class for USB transfer errors.

I had a lot of repeated status checking and transferList.remove code in
the transfer callbacks. I changed this to one callback that does this
for all the callbacks and then that calls the original callback.