Added logger for easier debugging
There's a lot of blanket exception catches in the code so if you have a
typo or something that causes an exception in some standard code, a lot
of time it will just get lost. This adds a logging thing that will log
all the exceptions to a file.

It's the standard Python logger so its also good for adding debug
logging statements that will only get printed if you set the DEBUG
environment variable to 1.